+++ auth_models.py
@@ -1,0 +1,78 @@
+from django.db import models
+from django.contrib.auth.models import User
+from django.conf import settings
+from typing import Optional
+import datetime
+from django.utils import timezone
+import qrcode
+import qrcode.constants
+import qrcode.image.svg
+import logging
+from io import BytesIO
+from django.utils.encoding import iri_to_uri
+from django.core.files import File
+
+import common_models.models as md
+logger = logging.getLogger("common_models.auth_models")
+
+
+class MagicLink(models.Model):
+    token = models.CharField(max_length=64, default=md.random_token)
+    user = models.OneToOneField(User, models.CASCADE)
+    expiry = models.DateTimeField(default=md.days5)
+    delete_immediately = models.BooleanField(default=True)
+
+    qr_code = models.ImageField(upload_to=md.magic_link_qr_code_path, blank=True)
+
+    def link_used(self) -> bool:
+        """Returns True if link can still be used, or False if not."""
+        if self.delete_immediately:
+            self.delete()
+            return False
+
+        if self.expiry < timezone.now() + datetime.timedelta(days=1):
+            # Only 1 day left, do nothing
+            return True
+
+        else:
+            self.expiry = timezone.now() + datetime.timedelta(days=1)
+            self.save()
+            return True
+
+    def full_link(
+            self, hostname: Optional[str] = None, login_path: Optional[str] = None, redirect: Optional[str] = None):
+
+        if hostname is None:
+            hostname_s = "https://" + settings.ALLOWED_HOSTS[0]
+        else:
+            hostname_s = hostname
+
+        DEFAULT_LOGIN_PATH = "/accounts/login"
+
+        if login_path is None:
+            login_path = DEFAULT_LOGIN_PATH
+
+        if redirect is not None:
+            redirect_str = f"&redirect={iri_to_uri(redirect)}"
+        else:
+            redirect_str = ""
+
+        if hostname_s[:8] != "https://" or hostname_s[:7] != "http://":
+            # TODO clean this up to make it better
+            hostname_s = "http://" + hostname_s
+
+        return f"{hostname_s}{login_path}?auth={self.token}{redirect_str}"
+
+    def _generate_qr_code(
+            self, hostname: Optional[str] = None, login_path: Optional[str] = None,
+            redirect: Optional[str] = None) -> None:
+
+        qr = qrcode.QRCode(error_correction=qrcode.constants.ERROR_CORRECT_H)
+        qr.add_data(self.full_link(hostname=hostname, login_path=login_path, redirect=redirect))
+        qr.make(fit=True)
+
+        blob = BytesIO()
+        img = qr.make_image()
+        img.save(blob, "PNG")
+        self.qr_code.save("QRCode.png", File(blob))
+        self.save()
